GUI cleanup, handling different windows
Handling multiple windows using GLUT added (in reality just
resizing/changing a single window). Each new window should have
windowName::show function to handle that.

All GUI classes separated into corresponding files. Might add a master
gui class so the window files don't have to import every object they're
using (can't put into gui.h due to redefinition problems)

Title now shows properly, now onto tidying it up.
