+++ source/client/gui/gui_defines.h
@@ -1,0 +1,10 @@
+#pragma once
+#define UI_MAX_NODES 50 //PER PANEL
+#define BTN_ACTIVE 0
+#define BTN_DOWN 1
+#define BTN_DISABLED 2
+#define BTN_HOVER 3
+
+#define COMBO_ACTIVE 0
+#define COMBO_OPEN 1
+#define COMBO_HOVER 4
