+++ source/client/options.cpp
@@ -42,40 +42,50 @@ bool g_savePassword;
 bool g_fullScreen;
 
 
+namespace options {
 
+	/*
+	*	Call this to display the options window
+	*/
+	void show() {
+		//loadSettings();
+		//loadAccData(); //if settings store userAcc data
+		windowSizeX = GetSystemMetrics(SM_CXSCREEN) / 5;
+		windowSizeY = GetSystemMetrics(SM_CYSCREEN) / 4 + 200;
+		glutReshapeWindow(windowSizeX, windowSizeY);
+		init();
+		glutDisplayFunc(display);
+		glutMouseFunc(mouseInput);
+		glutPassiveMotionFunc(mouseMove);
+		display();
+	}
 
+	/*
+	*	Frees memory of all the UI objects (static globals with 'ui_' prefix)
+	*/
+	void freeUI() {
+		delete ui_window;
+		delete ui_account;
+		delete ui_save;
+		delete ui_cancel;
 
-
-/*
-*	Initialising OpenGL inside Glut
-*	Author: Kamil
-*/
-static void init() {
-	glClearColor(0.0f / 255, 0.0f / 255, 0.0f / 255, 0.0);
-	glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
-	glEnable(GL_BLEND);
-	glShadeModel(GL_SMOOTH);
-	freeTitleUIObjects();
-	freeOptionsUIObjects(); //Free memory of UI items in case they already exist
-	ui_settings = new UI_Panel(0, 0, windowSizeX, windowSizeY);
-	ui_window = new UI_Panel(0, 0, windowSizeX, windowSizeY / 4 * 2);
-	ui_account = new UI_Panel(0, windowSizeY / 4 * 2, windowSizeX, windowSizeY / 4);
-	ui_buttons = new UI_Panel(0, windowSizeY / 4 * 3, windowSizeX, windowSizeY / 4);
-	ui_save = new UI_Button(*ui_buttons, 15, 10, windowSizeX / 9 * 3, 30, "Save");
-	ui_cancel = new UI_Button(*ui_buttons, 15 + windowSizeX / 2, 10, windowSizeX / 9 * 3, 30, "Cancel");
-	/*Set onClick actions*/
-	//ui_play->setAction(login);
-	//ui_options->setAction(options);
-	//ui_exit->setAction(exit);
-	//glMatrixMode(GL_MODELVIEW);
+	}
 
 }
+
+/*
+*	==========================
+*	==========================
+*	==   STATIC FUNCTIONS   ==
+*	==========================
+*	==========================
+*/
 
 /*
 *	All the drawing goes here
 *	Author: Kamil
 */
-void optionsDisplay(void) {
+static void display(void) {
 	printMsg(MSG_LOG, "Scene redrawn.");
 	bool windowChanged = false;
 	/*	Get current window sizes and change them accordingly
@@ -112,50 +122,90 @@ void optionsDisplay(void) {
 	glutSwapBuffers();
 }
 
-
 /*
-*	Initialising title window and glut
+*	Initialising Options
 *	Author: Kamil
 */
-static void window() {
-	/*Fake arguments to pass to glutInit()*/
-	int argc = 1;
-	char *argv[1];
-	argv[0] = "1";
-	/*Get resolution of the screen*/
-	int screenX = GetSystemMetrics(SM_CXSCREEN);
-	int screenY = GetSystemMetrics(SM_CYSCREEN);
-	/*Create a reasonably sized window*/
-	glutInitWindowSize(windowSizeX, windowSizeY);
-	glutInitWindowPosition(screenX / 2.8, screenY / 2 - windowSizeY / 1.5);
-	glutCreateWindow("Options");
-	init();
-	/*Redirect glut to specific functions*/
-	glutDisplayFunc(optionsDisplay);
-		//glutMouseFunc(mouseInput);
-		//glutPassiveMotionFunc(mouseMove);
-	optionsDisplay();
-	//glutIdleFunc(drawUI);
-	glutMainLoop();
+static void init() {
+	glClearColor(0.0f / 255, 0.0f / 255, 0.0f / 255, 0.0);
+	glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
+	glEnable(GL_BLEND);
+	glShadeModel(GL_SMOOTH);
+	options::freeUI(); //Free memory of UI items in case they already exist
+	ui_settings = new UI_Panel(0, 0, windowSizeX, windowSizeY);
+	ui_window = new UI_Panel(0, 0, windowSizeX, windowSizeY / 4 * 2);
+	ui_account = new UI_Panel(0, windowSizeY / 4 * 2, windowSizeX, windowSizeY / 4);
+	ui_buttons = new UI_Panel(0, windowSizeY / 4 * 3, windowSizeX, windowSizeY / 4);
+	ui_save = new UI_Button(*ui_buttons, 15, 10, windowSizeX / 9 * 3, 30, "Save");
+	ui_cancel = new UI_Button(*ui_buttons, 15 + windowSizeX / 2, 10, windowSizeX / 9 * 3, 30, "Cancel");
+	/*Set onClick actions*/
+	//ui_play->setAction(login);
+	//ui_options->setAction(options);
+	ui_cancel->setAction(cancel);
+	//glMatrixMode(GL_MODELVIEW);
 
 }
 
+
 /*
-*	Call this to display the options window
+*	Handles mouse input
 */
-void optionsCreate() {
-	//loadSettings();
-	//loadAccData(); //if settings store userAcc data
-	window();
+static void mouseInput(int key, int state, int x, int y) {
+	if (key == GLUT_LEFT_BUTTON) {
+		/* The following piece of code iterates over all the nodes in "ui_menu" object
+		and checks if they're within given X,Y coordinates. */
+
+		/* get nodes of the panel */
+		std::vector<UI_Node*>& nodes = ui_buttons->getNodes();
+		/* iterate over said nodes */
+		for (std::vector<UI_Node*>::iterator i = nodes.begin(); i != nodes.end(); ++i) {
+			/* check if the clicked area is within node's coordinates */
+			if (withinCoords(x, y, *i)) {
+				/* button event handling*/
+				UI_Button * btn = dynamic_cast<UI_Button*>(*i); //dynamic cast node->button
+				if (btn != nullptr) { //if cast was successfull (meaning this indeed is a button)
+					if (state == GLUT_DOWN) { //if mouseDown
+						btn->setState(BTN_DOWN);
+					}
+					else if (btn->getState() == BTN_DOWN) { //if mouse released
+						btn->setState(BTN_ACTIVE);
+						btn->onClick();
+					}
+					printMsg(MSG_LOG, "Clicked a button.");
+				}
+				/* Insert any other GUI elements handling under this comment */
+			}
+		}
+	}
+	glutPostRedisplay();
 }
 
 /*
-*	Frees memory of all the UI objects (static globals with 'ui_' prefix)
+*	Handles moving the mouse and hovering over objects
 */
-void freeOptionsUIObjects() {
-	free(ui_window);
-	free(ui_account);
-	free(ui_save);
-	free(ui_cancel);
+static void mouseMove(int x, int y) {
+	std::vector<UI_Node*>& nodes = ui_buttons->getNodes();
+	if (nodes.size() > 0) {
+		for (std::vector<UI_Node*>::iterator i = nodes.begin(); i != nodes.end(); ++i) {
+			UI_Button * btn = dynamic_cast<UI_Button*>(*i);
+			if (btn != nullptr) {
+				if (withinCoords(x, y, *i)) {
+					btn->setState(BTN_HOVER);
+				}
+				else {
+					if (btn->getState() != BTN_DISABLED) {
+						btn->setState(BTN_ACTIVE);
+					}
+				}
+			}
+		}
+	}
+	glutPostRedisplay();
+}
 
+/*
+*	Go back to main menu
+*/
+static void cancel() {
+	title::show();
 }
